Initial work in progress to support RISC OS extra fields.
The implementation inside JFPatch-as-a-service is rudimentary at
best (although actually sufficient for simple operations), but I
need to be able to both read information and write information.

Zipper provides some of the information needed for this to work,
with InfoZip and Python's own zip implementation being used for
some of the handling of flags.

It's not sufficient for use yet, but it's got some of the conversions
present.