Support for NFS encoding of the filename; tests for filetyping.
NFS encodings are now supported and should modify the base filename,
and keep the RISC OS attributes up to date. This should mean that we
get the right settings from the filenames, and that we can write to
them in the normal way.

Filetypes are now handled (hopefully) sensibly, with directory types
(&1000 in the API, &FFD in the load address) and untyped files (-1)
being handled sensibly, I think.