Added tests and fixes for attribute assignment; base tests improved.
Changing the attributes now changes the MSDOS and unix attributes
in the external attributes field. These are now exercised with and
without unix attributes set. It's likely that we will want to always
set the external attributes, but I'm not making that decision here.

The base tests now check some of the properties being updated in the
base, and their effects on the RISC OS properties.