Added support for Python 3.
The `riscos_filename` property is now a str-type object (bytes in
Python 3), whilst the `filename` object remains a unicode object.
This has meant some further qualifiers needed adding to the code
to make it work the same way on either version of Python.

A little more clarification added to the file prologue comment to
aid users in the use and intent of the variables.