Add from_file classmethod; improved nfs encoding support.
The from_file method is intended to take the details from a file on
disc and populate the right parameters. This isn't provided in the
Python 2 version of the ZipInfo API, but it is very useful in getting
the RISC OS information into a structure when using the nfs_extension
format on disc.

The nfs_encoding variable can now be turned on and off and will
update the filename which will be stored to include or exclude the
NFS encoding parameters for the files.