Fix for RISC OS extra fields on directories not setting loadaddr properly
The load address wasn't being set properly on the directories when they
were being read, resulting in 0s in the upper 24 bits of the load
address field. This seems to be caused by some archives writing those
bits as 0s. This is corrected on load, so that directories consistently
appear as files with the directory filetype.