Added sanitisation of the unix filenames.
The archive might contain names that are intended to subvert the host
system, using '..' or '.' or even rooted filenames. Even if not intended
for evil, these names will not be helpful within the RISC OS system.

A 'sanitise_unix' method has been introduced which attempts to reduce
these names to something that is suitable for use on RISC OS.