Make __repr__ safer if invalid values are used.
If invalid values are used the __repr__could give errors whilst
trying to present the results. These are now trapped and errors
reported rather than being raised. This probably only happens if
the datetime is invalid, in which case we probably should trap
it earlier, but for now this at least prevents the reporting of
the object from failing.