Updated the zip creation system to be RISC OS.
We're creating in the RISC OS form as a RISC OS tool, so we now set
the create_system, to match how the RISC OS tools work. It doesn't
seem like anyone trusts these values (other than the Amiga) so it's
not actually that important, but it's at least consistent.