Add rozipfile module, and command line tool.
The `rozipfile` module is similar to the `zipfile` module and can
extract or create archives for RISC OS files. I've created it largely
to allow me to import the module so that I can use it to construct
release zip archives in RISC OS format, instead of the NFS encoded
format I've been using, through a simple command.