Fix the rozipfile invocation in Python 3.
When run in Python 3 there are some problems with the bytes objects
for filenames which mean that it's not able to sort and gives some
exceptions. This and other problems related to Python 3 were addressed.

A formatting issue which produced a rogue '%' on some Python versions
in the date format was fixed, too.