Improved handling of the encoding of filenames.
The filenames handling was not working properly when the zip archive
was a RISC OS format archive using its own encoding. This was made
harder by the manner in which the ZipInfo object can be created
(either as part of a zip file being read, or manually, or through
the creation from a file on disc), and the version of Python that
is in use. We now settle on `filename` being unicode, and
`riscos_filename` being the representation on RISC OS of that name.

This still doesn't address the problem that the separator is going
to be mangled by the standard ZipInfo implementation when used on
RISC OS. There's little I can do about that at this second.