Ensure that directories are created implicitly on extract.
When we extract files we usually expect the directories to be present
before we try writing to the files. However, it's entirely possible
to have file entries appear before the directory entries - possible but
odd. As it can happen, we now ensure that the directories are created.
This allows the RISCOSM archives to be extracted, which exhibit this
feature.