Add support for default filetype; improved testing of rozipfile.
The default filetype can now be specified to the tools, and this is
now tested by the test code for showzip.

There is now some simple test code for rozipfile which exercises the
listing interfaces. This found some issues on RISC OS 3 with the
encoding not coming out right. I think this has now been fixed.