Fix for directories ending in '/' without MSDOS Dir attribute.
One of the very old archives I encountered had directories that ended
with '/' but which didn't have the msdos_directory attribute bit set.
This meant that we created files with the name of the directory and
then couldn't create the files underneath them.

We now implicitly flag the directory if it ends in '/'.