Add support for compression levels; default to compressing.
The zipfile module now supports compression levels, if the underlying
Python zipfile module supports them. This means that with Python 3.7
and higher we get control of the compression levels. On earlier Python
versions, the compression is limited to the choice between the
default of compressing with deflate, or storing.

Previously we only ever stored the files. We now support compressing
the files.